Fix multiple bugs: null pointer deref, integer overflow, resource safety
- x11_canvas.cpp: Fix measureText null pointer deref when display_ is null
  (early return before XftFontOpenName call); clear backbuffer on resize
  to prevent uninitialized garbage on first paint; remove beginPaint
  backbuffer clear to preserve dirty-rect painting correctness.

- x11_window.cpp: Add missing <X11/Xatom.h> and <unistd.h> includes.
  Fix None macro usage (replaced with 0) since None is #undef'd after
  X11 header inclusion. Reorder namespace declaration after system includes.

- scrollarea.cpp: Fix integer overflow when height >= 46341
  (height * height overflows int32). Use int64_t casts in all three
  scrollbar thumb height calculations and scroll position computations.

- tableview.cpp: Fix integer overflow in row index calculation
  (scrollY_ * rowHeight_ overflows for large tables).

- win32_canvas.h, x11_canvas.h, cocoa_canvas.h: Delete copy
  constructor/assignment to prevent accidental double-free of
  platform resources (GDI+, X11, CGContext/ObjC).

+++ src/widgets/scrollarea.cpp
@@ -82,11 +82,12 @@ void ScrollArea::paintSelf(NativeCanvas* canvas) {
     if (contentHeight_ > 0 && contentHeight_ > height()) {
         int sbWidth = 12;
         int scrollY = currentScrollY();
-        int thumbH = std::max(24, height() * height() / contentHeight_);
+        // Use int64_t to avoid overflow when height >= 46341 (sqrt(INT_MAX))
+        int thumbH = std::max(24, (int)((int64_t)height() * height() / contentHeight_));
         int maxScroll = contentHeight_ - height();
         int thumbY = r.y;
         if (maxScroll > 0) {
-            thumbY += (scrollY * (height() - thumbH)) / maxScroll;
+            thumbY += (int)((int64_t)scrollY * (height() - thumbH) / maxScroll);
         }
 
         canvas->setColor(t.scrollbarTrack);
@@ -121,7 +122,7 @@ bool ScrollArea::handleEvent(Event& event) {
     if (event.type == EventType::MouseDown && onScrollbar) {
         // Calculate scrollbar thumb position (contentHeight_ > 0 guaranteed by onScrollbar)
         int curScrollY = currentScrollY();
-        int thumbH = std::max(24, (height() * height()) / contentHeight_);
+        int thumbH = std::max(24, (int)((int64_t)height() * height() / contentHeight_));
         int maxScroll = contentHeight_ - height();
         int thumbY = 0;
         if (maxScroll > 0) {
@@ -146,7 +147,7 @@ bool ScrollArea::handleEvent(Event& event) {
     }
 
     if (event.type == EventType::MouseMove && draggingScrollbar_) {
-        int thumbH = std::max(24, height() * height() / contentHeight_);
+        int thumbH = std::max(24, (int)((int64_t)height() * height() / contentHeight_));
         int maxScroll = contentHeight_ - height();
         int trackH = height() - thumbH;
         if (maxScroll > 0 && trackH > 0) {
